refactor: move Rating to folder + cleanup old files + fix molecules index
