+++ src/components/molecules/index.ts
@@ -1,3 +1,2 @@
-export { Card } from './Card'
-export { ProductCard } from './ProductCard'
-export { Rating } from './Rating'
+export { Card, ProductCard } from './Card/Card'
+export { Rating } from './Rating/Rating'
