--- src/components/molecules/index.tsx
@@ -1,3 +1,0 @@
-// src/components/molecules/index.ts
-export { Rating } from './Rating'
-export { Card, ProductCard } from './Card'
